+++ src/pages/passport/Layout.astro
@@ -7,7 +7,7 @@ import '@devprotocol/clubs-core/styles'
 import type { Skin } from '@pages/api/profile'
 import Header from '@components/Global/Header.astro'
 import { getPassportItemForPayload } from '@fixtures/api/passportItem'
-import SkinSwitch from '@components/Passports/SkinSwitch.svelte'
+import SkinSwitch from 'src/pages/passport/components/SkinSwitch.svelte'
 
 type Props = {
   eoa: string
