+++ src/pages/passport/[id]/clubs.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '@layouts/Landing.astro'
 
-import UserClubs from '@components/Passports/UserClubs.svelte'
+import UserClubs from '../components/UserClubs.svelte'
 
 const { id } = Astro.params
 ---
