+++ src/pages/passport/components/IconSpotlight.svelte
