+++ src/pages/passport/components/IconXMark.svelte
