+++ src/pages/passport/components/PassportAsset.svelte
