+++ src/pages/passport/components/PassportClipEditModal.svelte
