+++ src/pages/passport/components/SkinSwitch.svelte
