+++ src/pages/passport/components/UserClubItem.svelte
